requirements.txt
qrcode~=7.3.1
robonomics-interface
matplotlib~=3.5.1
Pillow~=9.0.1